3 changes: `Set ID → IntSet`; better CLI/tests; better (?) `equiv`
`equiv`, specifically `transitions`, was by FAR the slowest part of the app,
which was having what seemed to be >O(n log n) running time.

Now that I've gone to test the changes, though, I'm seeing a very occasional
test failure.  Unfortunately, the way the test was written, I'm on "178.195
shrinks" and counting, so I don't know what it was yet…